@@ -15,8 +15,8 @@ android {
     }
     compileOptions {
         // See : https://developer.android.com/studio/write/java8-support.html
-        sourceCompatibility JavaVersion.VERSION_1_7
-        targetCompatibility JavaVersion.VERSION_1_7
+        sourceCompatibility JavaVersion.VERSION_1_8
+        targetCompatibility JavaVersion.VERSION_1_8
     }
     signingConfigs {
     }
